+++ ChangeLog.md
@@ -4,6 +4,7 @@ ChangeLog
 ## unreleased:
 * Add `{{symbol}}` to get a new replacement for a symbol.
 * Add new escapement '~ ' to escape spaces.
+* Add `instantiate_from_seed` to allow for deterministic result.
 
 ## 0.1.0 (2019-04-13):
 * Add support for JSON.